requirements.txt
psycopg2-binary==2.9.6
python-dateutil==2.8.2
six==1.16.0
